factor out equiv from symbolic rewriter
